+++ private/Test-LinkedServiceConnection.ps1
@@ -2,7 +2,11 @@
 # https://datathirst.net/blog/2018/9/23/adfv2-testing-linked-services
 
 
-function Get-Bearer([string]$TenantID, [string]$ClientID, [string]$ClientSecret)
+function Get-Bearer(
+  [string] $TenantID, 
+  [string] $ClientID, 
+  [string] $ClientSecret
+)
 {
   $TokenEndpoint = {https://login.windows.net/{0}/oauth2/token} -f $TenantID 
   $ARMResource = "https://management.core.windows.net/";
@@ -14,7 +18,7 @@ function Get-Bearer([string]$TenantID, [string]$ClientID, [string]$ClientSecret)
           'client_secret' = $ClientSecret
   }
 
-  $params = @{
+  $request = @{
       ContentType = 'application/x-www-form-urlencoded'
       Headers = @{'accept'='application/json'}
       Body = $Body
@@ -22,41 +26,52 @@ function Get-Bearer([string]$TenantID, [string]$ClientID, [string]$ClientSecret)
       URI = $TokenEndpoint
   }
 
-  $token = Invoke-RestMethod @params
+  $token = Invoke-RestMethod @request
 
   return "Bearer " + ($token.access_token).ToString()
 }
 
 
-function Get-LinkedServiceBody([string]$LinkedServiceName, [string]$DataFactoryName, [string]$ResourceGroupName, [string]$BearerToken, [string]$SubscriptionID)
+function Get-LinkedServiceBody(
+  [string] $LinkedServiceName, 
+  [string] $DataFactoryName, 
+  [string] $ResourceGroupName, 
+  [string] $BearerToken, 
+  [string] $SubscriptionID
+)
 {
   Write-Debug "BEGIN: Get-LinkedServiceBody()"
   $ADFEndpoint = "https://management.azure.com/subscriptions/$SubscriptionId/resourceGroups/$ResourceGroupName/providers/Microsoft.DataFactory/factories/$DataFactoryName/linkedservices/$($LinkedServiceName)?api-version=2018-06-01"
 
-  $params = @{
+  $request = @{
       ContentType = 'application/json'
       Headers = @{'accept'='application/json';'Authorization'=$BearerToken}
       Method = 'GET'
       URI = $ADFEndpoint
   }
 
-  $a = Invoke-RestMethod @params
+  $a = Invoke-RestMethod @request
   Write-Debug "END: Get-LinkedServiceBody()"
   return ConvertTo-Json -InputObject @{"linkedService" = $a} -Depth 50
 }
 
-function Get-LinkedServiceBodyAzRestMethod([string]$LinkedServiceName, [string]$DataFactoryName, [string]$ResourceGroupName, [string]$SubscriptionID)
+function Get-LinkedServiceBodyAzRestMethod(
+  [string] $LinkedServiceName, 
+  [string] $DataFactoryName, 
+  [string] $ResourceGroupName, 
+  [string] $SubscriptionID
+)
 {
   Write-Debug "BEGIN: Get-LinkedServiceBodyAzRestMethod()"
 
   $ADFEndpoint = "/subscriptions/$SubscriptionId/resourceGroups/$ResourceGroupName/providers/Microsoft.DataFactory/factories/$DataFactoryName/linkedservices/$($LinkedServiceName)?api-version=2018-06-01"
 
-  $params = @{
+  $request = @{
       Path = $ADFEndpoint
       Method = 'GET'
   }
 
-  $a = Invoke-AzRestMethod @params
+  $a = Invoke-AzRestMethod @request
   $a = ($a.Content | ConvertFrom-Json)
   #Write-Debug (ConvertTo-Json -InputObject @{"linkedService" = $a} -Depth 50)
   Write-Debug "END: Get-LinkedServiceBodyAzRestMethod()"
@@ -64,7 +79,14 @@ function Get-LinkedServiceBodyAzRestMethod([string]$LinkedServiceName, [string]$
 }
 
 
-function Test-LinkedServiceConnection([string]$LinkedServiceName, [string]$DataFactoryName, [string]$ResourceGroupName, [string]$BearerToken, [string]$SubscriptionID)
+function Test-LinkedServiceConnection(
+  [string] $LinkedServiceName, 
+  [string] $DataFactoryName, 
+  [string] $ResourceGroupName, 
+  [string] $BearerToken, 
+  [string] $SubscriptionID,
+  $Params
+)
 {
   Write-Debug "BEGIN: Test-LinkedServiceConnection()"
 
@@ -72,16 +94,27 @@ function Test-LinkedServiceConnection([string]$LinkedServiceName, [string]$DataF
 
   $AzureEndpoint = "https://management.azure.com/subscriptions/$SubscriptionID/resourcegroups/$ResourceGroupName/providers/Microsoft.DataFactory/factories/$DataFactoryName/testConnectivity?api-version=2018-06-01"
 
-  $params = @{
+  if ($Params) {
+    Write-Verbose "Found input parameters for this Linked Service, adding to the API request..."
+    $j = ConvertFrom-Json $body 
+    $j.linkedService.properties.parameters = $Params
+    $body = ConvertTo-Json -InputObject $j -Depth 50
+  }
+
+  Write-Verbose '============== Linked Service to being connection tested ============='
+  Write-Verbose $body
+  Write-Verbose '======================================================================'
+  
+  $request = @{
       ContentType = 'application/json'
       Headers = @{'accept'='application/json';'Authorization'=$BearerToken}
       Body = $Body
-      Method = 'Post'
+      Method = 'POST'
       Uri = $AzureEndpoint
   }
 
   try {
-    $response = Invoke-RestMethod @params
+    $response = Invoke-RestMethod @request
   }
   catch {
     Write-Error -Exception $_.Exception
@@ -90,7 +123,12 @@ function Test-LinkedServiceConnection([string]$LinkedServiceName, [string]$DataF
   return $response
 }
 
-function Test-LinkedServiceConnectionAzRestMethod([string]$LinkedServiceName, [string]$DataFactoryName, [string]$ResourceGroupName, [string]$SubscriptionID)
+function Test-LinkedServiceConnectionAzRestMethod(
+  [string] $LinkedServiceName, 
+  [string] $DataFactoryName, 
+  [string] $ResourceGroupName, 
+  [string] $SubscriptionID
+)
 {
   Write-Debug "BEGIN: Test-LinkedServiceConnectionAzRestMethod()"